% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppColMetric-package.R
\docType{package}
\name{RcppColMetric-package}
\alias{RcppColMetric}
\alias{RcppColMetric-package}
\title{RcppColMetric: Efficient Column-Wise Metric Computation Against Common Vector}
\description{
In data science, it is a common practice to compute a series of columns (e.g. features) against a common response vector. Various metrics are provided with efficient computation implemented with 'Rcpp'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zhuxr11/RcppColMetric}
  \item Report bugs at \url{https://github.com/zhuxr11/RcppColMetric/issues}
}

}
\author{
\strong{Maintainer}: Xiurui Zhu \email{zxr6@163.com}

}
\keyword{internal}
